Moved from "Root Dir" to "Project Dir"
Side effect is that there is now an `app/` folder that mimics
Symfony Standard. Upshot is that `cache` and `logs` now live
under `app/`, though I'm not sure this is actually a pro.

Main benefit is that this will look and feel more like a
Symfony Standard application and since we are already
sorta going for that I think this was the right move.

The new app/configuration stuff are as follows:

 * app/SculpinKernel.php
 * app/config/sculpin_kernel.yml
 * app/config/sculpin_site.yml

The sculpin_kernel.yml is the configuration file for the
Sculpin Kernel. (surprise!) This is located here in the
event that a standalone Sculpin Kernel exists inside of
a Symfony Standard application.

The sculpin_site.yml is the site configuration used as
the `site.*` values in the templates. This is almost
entirely meta data for the time being. This file is
read every time Sculpin fires up and is not currently
cached in any way.