Import original data for a source when init is called again
There exists cases where calling init on an existing source would lose data set
from previous calls. The problem is that any time a file is asked to be
reprocessed (happens when updated as a part of a collection and other hard to
track down cases) the metadata would be read in again from the file. The side
effect of this is that ONLY the metadata coming in from the frontmatter would
remain. Any metadata set by other extensions or other parts of Sculpin core
(things like date tracking and auto layout for proxy source collections) would
be lost on init.

This patch works by saving a copy of the data before the new metadata is
processed. Afterwards, if data was there originally, it imports it on top of the
new metdata w/ clobber set to `false`. This will ensure that data that is
already set will not be overwritten by the old data.